v6

+++ CBayMobileApp/ViewModels/Shopping/CartOrderConfirmationViewModel.cs
@@ -125,8 +125,8 @@ namespace CBayMobileApp.ViewModels.Shopping
             }
         }
 
-        private int availableBalance;
-        public int AvailableBalance
+        private double  availableBalance;
+        public double  AvailableBalance
         {
             get => availableBalance;
             set
